Teach destroy_repository.py to destroy the filenames of the given RCS files
This patch is somewhat complicated by the need to rewrite CVS log messages
of the form "file X was initially added on branch Y". Rewriting these log
messages is still done, even if metadata is not to be destroyed.

Filenames are rewritten with diregard to any leading directories and
trailing ',v'-extensions, which means that the "same" filename in various
forms will be rewritten in the same manner. For example:
	foo.txt             -> file001
	bar.txt             -> file002
	bar.txt,v           -> file002,v
	baz/bar.txt         -> baz/file002
	/home/baz/bar.txt,v -> /home/baz/file002,v

Patch by: Johan Herland <johan@herland.net>


git-svn-id: http://cvs2svn.tigris.org/svn/cvs2svn/trunk@4609 be7e6eca-30d4-0310-a8e5-ac0d63af7087