destroy_repository.py: Make the "destroyed" text more useful.
Now that the empty/non-empty state of revisions is preserved by
destroy_repository.py, non-empty revisions necessarily have some
contents.  So let's make those contents a little bit useful for
debugging.

Trunk and branch revisions are treated differently because if trunk
revision N was preceded by a revision N-1 with an empty deltatext,
then the text that appears in the delta for revision N was actually
committed in N-1.

git-svn-id: http://cvs2svn.tigris.org/svn/cvs2svn/trunk@5067 be7e6eca-30d4-0310-a8e5-ac0d63af7087