Replace algorithm in SubtreeSymbolTransform.
Patch by: Jon Foster <jon.foster@cabot.co.uk>

The algorithm used by SubtreeSymbolTransform.__does_rule_apply_to()
was written to be simple and similar to the existing logic in
SubtreeSymbolMapper.transform().  Unfortunately, profiling shows that
it's extremely slow.  So it's time to replace it with an algorithm
written for speed.

git-svn-id: http://cvs2svn.tigris.org/svn/cvs2svn/trunk@5110 be7e6eca-30d4-0310-a8e5-ac0d63af7087