Change the representation of "blocks" within the rcs_stream module.
Represent blocks as 'c'opies and 'r'eplaces, where the replace blocks
can contain both old_lines and new_lines.  Don't bother keeping track
of line numbers, as (with this format) they are only needed on output.
This makes it trivial to implement transformations on blocks.  The
cost is that sometimes "old_lines" is generated when it is not needed.
But it only consists of a list of pointers to lines that are already
in RAM, so it is not much extra work.

git-svn-id: http://cvs2svn.tigris.org/svn/cvs2svn/trunk@5128 be7e6eca-30d4-0310-a8e5-ac0d63af7087